move search and search functionality into the screen and pass the result to the users modal for rendering
