+++ app/screens/create_direct_message/create_direct_message.tsx
@@ -1,17 +1,32 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
-import React, {useCallback, useState} from 'react';
+import React, {useCallback, useRef, useState} from 'react';
 import {defineMessages, useIntl} from 'react-intl';
+import {Platform, SafeAreaView, StyleSheet, View} from 'react-native';
 
 import {makeDirectChannel, makeGroupChannel} from '@actions/remote/channel';
 import {fetchProfiles, fetchProfilesInTeam, searchProfiles} from '@actions/remote/user';
+import Search from '@components/search';
 import UsersModal from '@components/users_modal';
 import {General} from '@constants';
 import {useServerUrl} from '@context/server';
+import {useTheme} from '@context/theme';
 import {t} from '@i18n';
 import {alertErrorWithFallback} from '@utils/draft';
+import {changeOpacity, getKeyboardAppearanceFromTheme} from '@utils/theme';
 import {displayUsername} from '@utils/user';
+
+const style = StyleSheet.create({
+    container: {
+        flex: 1,
+    },
+    searchBar: {
+        marginLeft: 12,
+        marginRight: Platform.select({ios: 4, default: 12}),
+        marginVertical: 12,
+    },
+});
 
 const messages = defineMessages({
     dm: {
@@ -44,10 +59,16 @@ export default function CreateDirectMessage({
     teammateNameDisplay,
 }: Props) {
     const serverUrl = useServerUrl();
+    const theme = useTheme();
     const intl = useIntl();
 
     const [selectedIds, setSelectedIds] = useState<{[id: string]: UserProfile}>({});
-    const [page, setPage] = useState(-1);
+    const [searchResults, setSearchResults] = useState<UserProfile[]>([]);
+    const [term, setTerm] = useState('');
+    const [loading, setLoading] = useState(false);
+
+    const pageRef = useRef<number>(-1);
+    const searchTimeoutId = useRef<NodeJS.Timeout | null>(null);
 
     const createDirectChannel = useCallback(async (id: string): Promise<boolean> => {
         const user = selectedIds[id];
@@ -70,9 +91,9 @@ export default function CreateDirectMessage({
 
     const getProfiles = useCallback(async () => {
         if (restrictDirectMessage) {
-            return fetchProfilesInTeam(serverUrl, currentTeamId, page + 1, General.PROFILE_CHUNK_SIZE);
+            return fetchProfilesInTeam(serverUrl, currentTeamId, pageRef.current + 1, General.PROFILE_CHUNK_SIZE);
         }
-        return fetchProfiles(serverUrl, page + 1, General.PROFILE_CHUNK_SIZE);
+        return fetchProfiles(serverUrl, pageRef.current + 1, General.PROFILE_CHUNK_SIZE);
     }, [restrictDirectMessage, serverUrl, currentTeamId]);
 
     const onButtonTap = useCallback(async (selectedId?: {[id: string]: boolean}) => {
@@ -91,21 +112,76 @@ export default function CreateDirectMessage({
         return searchProfiles(serverUrl, lowerCasedTerm, {allow_inactive: true});
     }, [restrictDirectMessage, serverUrl, currentTeamId]);
 
+    const handleSearchUsers = useCallback(async (searchTerm: string) => {
+        setLoading(true);
+
+        const results = await searchUsers(searchTerm);
+
+        setSearchResults(results?.data || []);
+        setLoading(false);
+    }, [searchUsers]);
+
+    const search = useCallback(() => {
+        handleSearchUsers(term);
+    }, [handleSearchUsers, term]);
+
+    const clearSearch = useCallback(() => {
+        setTerm('');
+        setSearchResults([]);
+    }, []);
+
+    const onSearch = useCallback((text: string) => {
+        if (text) {
+            setTerm(text);
+            if (searchTimeoutId.current) {
+                clearTimeout(searchTimeoutId.current);
+            }
+
+            searchTimeoutId.current = setTimeout(() => {
+                handleSearchUsers(text);
+            }, General.SEARCH_TIMEOUT_MILLISECONDS);
+            return;
+        }
+
+        clearSearch();
+    }, [clearSearch, handleSearchUsers]);
+
     return (
-        <UsersModal
-            buttonText={messages.button}
-            componentId={componentId}
-            currentUserId={currentUserId}
-            getProfiles={getProfiles}
-            maxSelectedUsers={General.MAX_USERS_IN_GM}
-            page={page}
-            searchUsers={searchUsers}
-            selectedIds={selectedIds}
-            setPage={setPage}
-            setSelectedIds={setSelectedIds}
-            teammateNameDisplay={teammateNameDisplay}
-            onButtonTap={onButtonTap}
-        />
+        <SafeAreaView
+            style={style.container}
+            testID='members_modal.screen'
+        >
+            <View style={style.searchBar}>
+                <Search
+                    testID='members_modal.search_bar'
+                    placeholder={intl.formatMessage({id: 'search_bar.search', defaultMessage: 'Search'})}
+                    cancelButtonTitle={intl.formatMessage({id: 'mobile.post.cancel', defaultMessage: 'Cancel'})}
+                    placeholderTextColor={changeOpacity(theme.centerChannelColor, 0.5)}
+                    onChangeText={onSearch}
+                    onSubmitEditing={search}
+                    onCancel={clearSearch}
+                    autoCapitalize='none'
+                    keyboardAppearance={getKeyboardAppearanceFromTheme(theme)}
+                    value={term}
+                />
+            </View>
+            <UsersModal
+                ref={pageRef}
+                term={term}
+                clearSearch={clearSearch}
+                buttonText={messages.button}
+                componentId={componentId}
+                currentUserId={currentUserId}
+                getProfiles={getProfiles}
+                loading={loading}
+                setLoading={setLoading}
+                maxSelectedUsers={General.MAX_USERS_IN_GM}
+                searchResults={searchResults}
+                selectedIds={selectedIds}
+                setSelectedIds={setSelectedIds}
+                teammateNameDisplay={teammateNameDisplay}
+                onButtonTap={onButtonTap}
+            />
+        </SafeAreaView>
     );
 }
-
